+++ CHANGELOG.md
@@ -1,6 +1,30 @@
 Changelog
 =========
 
+2.1.0
+-----
+
+### Added
+
+* Individual rules in the `cache_control` can now again have a `match_response`
+  or `additional_response_status` configuration to limit the rule to certain
+  responses.
+
+  For this, the signature of CacheControlListener::addRule had to be changed.
+  It now expects a RuleMatcherInterface instead of the
+  ResponseMatcherInterface. If you extended the listener or change the service
+  configuration, this could be a **BC BREAK** for your application.
+
+### Fixed
+
+* If no response matching is configured on `cache_control`, the global
+  `cacheable` configuration is now respected to decide whether cache headers
+  should be set. By default, this follows RFC 7234, only responses with status
+  200, 203, 204, 206, 300,  301, 404, 405, 410, 414 or 501 get cache headers.
+
+  We decided to consider this a bugfix, but if your relied on this behaviour it
+  will be a **BC BREAK** for your application.
+
 2.0.0
 -----
 
@@ -27,7 +51,9 @@ Changelog
 ### Rule matcher
 
 * **BC break:** The `match_response` and `additional_cacheable_status` 
-  configuration parameters were removed for individual match rules. 
+  configuration parameters were removed for individual match rules.
+  Note that `match_response` and `additional_response_status` have been re-added for
+  `cache_control` in 2.0.1.
 
 * **BC break:** The second argument of the `RuleMatcher` constructor was changed 
   to take a `ResponseMatcherInterface`.